Improves WiFi configuration and reset features.
Enhances the WiFi configuration portal by improving the user
experience with automatic page reloads after saving settings
and performing a factory reset.

Moves the factory reset function to the footer for better
organization and adds a margin to the footer for improved layout.
Removes a link to the device status page.

Adds access to the underlying AsyncWebServer instance.